Ugly, works for simple cases
WIP: this works for the most basic case of including a whole package.
TODO: support sub-packages, hashes, tags, etc.
