+++ compiler/quilt/tools/const.py
@@ -26,6 +26,7 @@ RESERVED = {
     'environments': 'environments',
     'file': 'file',
     'kwargs': 'kwargs',
+    'package' : 'package',
     'transform': 'transform'
 }
