+++ compiler/quilt/tools/package.py
@@ -202,6 +202,22 @@ class Package(object):
             if not os.path.exists(path):
                 raise PackageException("Missing object fragments; re-install the package")
 
+    def save_package_tree(self, fullname, pkgnode):
+        """
+        Adds a package or sub-package tree from an existing package to this package's
+        contents.
+        """
+        contents = self.get_contents()
+        ipath = fullname.split('.')
+        leaf = ipath.pop()
+
+        ptr = contents
+        for node in ipath:
+            ptr = ptr.children.setdefault(node, GroupNode(dict()))
+        
+        ptr.children[leaf] = pkgnode
+
+
     def save_cached_df(self, hashes, name, path, ext, target, fmt):
         """
         Save a DataFrame to the store.
